Fix seed: ext (extensão de rua) é fracionário → double, não int
O seed de `street` estourava `System.FormatException` no GetInt32() porque
255 das 4426 ruas têm `ext` fracionário (ex.: 66.65 m) — assim como `larg`,
que já era double. municipio populava OK; street abortava o sidecar (exit 134).

- Street.Ext: int? → double? (modela metros fracionários; consistente com Larg).
- Seeder: ext lido com Dbl() em vez de Int().
- schema.sql: street.ext → double precision + ALTER idempotente (a tabela em
  prod foi criada como integer pelo initdb; o sidecar psql aplica o widening).
- Seeder.Int(): agora tolerante (TryGetInt32 → senão arredonda double → senão
  null) como rede de segurança contra números fora-de-int32 em campos futuros.

Verificado: ceps.json está limpo (idx int, lat/lng float, 6 cols) — sem risco
do mesmo crash no próximo estágio.

+++ src/TheDecrypter.Domain/Entities/Street.cs
@@ -10,6 +10,6 @@ public class Street
     public int? NumLei { get; set; }
     public string? DataLei { get; set; }
     public string? Localizacao { get; set; }
-    public int? Ext { get; set; }
+    public double? Ext { get; set; }   // extensão em metros (pode ser fracionária)
     public double? Larg { get; set; }
 }
